fix(R.7): close qa fallback blockers
